refactor(m9): fusionner AuthContext + ProfileContext en UserContext #264
- Absorbe toute la logique de ProfileContext dans AuthContext (setDisplayName
  avec vérif de conflit, refreshProfile, profileLoading state)
- Supprime le double fetch Supabase au login : un seul fetchProfile appelé
  via onAuthStateChange / getSession, protégé par profileLoadingRef
- Exporte useProfile() comme alias rétro-compatible depuis AuthContext
  (retourne { profile, loading, setDisplayName, refreshProfile })
- Supprime ProfileContext.tsx et retire ProfileProvider de App.tsx
- Met à jour les 2 consommateurs : Profile.tsx et UsernameRequiredDialog.tsx
- Build OK (exit 0), 169 tests passés
